refactor(st-09049): modularize tool registry
